Computational Biology/Salmonella outbreak/requirements.txt
biopython
matplotlib
numpy==1.19.3
